arm64: init the memory system before console/xenbus/gic
The mappings for console/xenbus/gic need several pages to setup
the page tables.

So we init the memory system before initiating the console/xenbus/gic,
and then we can allocate the pages which are needed by
the console/xenbus/gic mapping.

we use the 2M memory block for the memory.

Change-Id: I0d075478c107cf680d1a20989d57c0fcd727ab29
Jira: ENTOS-247
Signed-off-by: Huang Shijie <shijie.huang@arm.com>