mini-os: add the libfdt library
Add the libfdt libfdt which is BSD licence, it will used by
the arm code for the Flat Device Tree manipulation.

This libfdt is copied from:
       git://git.kernel.org/pub/scm/utils/dtc/dtc.git

This patch also changes makefile, makes it to compile the
libfdt library for arm64 code.

Also update the .gitignore, do not show some files for arm64.

Signed-off-by: Huang Shijie <shijie.huang@arm.com>